refactor: move middleware and utility functions into respective packages
